rename `tokenzier.ts` to `tokenizer.ts`
