+++ packages/diffs/test/editorBracketMatch.test.ts
@@ -5,7 +5,7 @@ import { DEFAULT_THEMES } from '../src/constants';
 import { Editor, type EditorOptions } from '../src/editor/editor';
 import { findBracketMatchRanges } from '../src/editor/matchBrackets';
 import { TextDocument } from '../src/editor/textDocument';
-import { EditorTokenizer } from '../src/editor/tokenzier';
+import { EditorTokenizer } from '../src/editor/tokenizer';
 import { disposeHighlighter } from '../src/highlighter/shared_highlighter';
 import type { FileContents } from '../src/types';
 import { installDom, wait } from './domHarness';
